% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{JMatrixSetDebug}
\alias{JMatrixSetDebug}
\title{JMatrixSetDebug}
\usage{
JMatrixSetDebug(deb = TRUE)
}
\arguments{
\item{deb}{boolean, TRUE to generate debug messages and FALSE to turn them off. Default: true}
}
\value{
No return value, called for side effects (internal boolean flag changed)
}
\description{
Sets debugging in jmatrix package to ON (with TRUE) or OFF (with FALSE).\cr
On package load the default status is OFF.\cr
Setting debugging to ON shows a message. Setting to OFF does not show anything (since debugging is OFF...)
}
\examples{
JMatrixSetDebug(TRUE)
JMatrixSetDebug(FALSE)
}
